+++ R/as.speeches.R
@@ -1,4 +1,4 @@
-#' @include S4classes.R make_region_matrix.R
+#' @include S4classes.R
 NULL
 
 #' Split corpus or partition into speeches.
@@ -26,6 +26,7 @@ NULL
 #'   \code{mc} is passed into \code{mclapply} as argument \code{mc.cores}.
 #' @param verbose A \code{logical} value, defaults to \code{TRUE}.
 #' @param progress A \code{logical} value, whether to show progress bar.
+#' @param xml Whether XML is "flat" or "nested" (length-one `character` vector).
 #' @param ... Further arguments.
 #' @return A \code{partition_bundle}, the names of the objects in the bundle are
 #'   the speaker name, the date of the speech and an index for the number of the
@@ -53,7 +54,7 @@ setMethod("as.speeches", "partition", function(
   .Object,
   s_attribute_date = grep("date", s_attributes(.Object), value = TRUE),
   s_attribute_name = grep("name", s_attributes(.Object), value = TRUE),
-  gap = 500, mc = FALSE, verbose = TRUE, progress = TRUE
+  gap = 500, xml = "flat", mc = FALSE, verbose = TRUE, progress = TRUE
 ){
   
   stopifnot(
@@ -83,7 +84,8 @@ setMethod("as.speeches", "partition", function(
         partition_bundle_names <- partition(
           partition_date,
           def = setNames(list(speaker_name), s_attribute_name),
-          verbose = FALSE
+          verbose = FALSE,
+          xml = xml
         )
         split(partition_bundle_names, gap = gap, verbose = FALSE)
       }
@@ -142,13 +144,13 @@ setMethod("as.speeches", "subcorpus", function(
   .Object,
   s_attribute_date = grep("date", s_attributes(.Object), value = TRUE),
   s_attribute_name = grep("name", s_attributes(.Object), value = TRUE),
-  gap = 500, mc = FALSE, verbose = TRUE, progress = TRUE
+  gap = 500, xml = "flat", mc = FALSE, verbose = TRUE, progress = TRUE
 ){
   as.speeches(
     .Object = as(.Object, "partition"),
     s_attribute_date = s_attribute_date,
     s_attribute_name = s_attribute_name,
-    gap = gap,
+    gap = gap, xml = xml,
     mc = mc, verbose = verbose, progress = progress
   )
 }
@@ -158,25 +160,45 @@ setMethod("as.speeches", "subcorpus", function(
 #' @rdname as.speeches
 #' @examples
 #' sp <- as.speeches(.Object = corpus("GERMAPARLMINI"), s_attribute_name = "speaker")
+#' @importFrom RcppCWB s_attr_regions s_attr_is_sibling s_attr_is_descendent
 setMethod("as.speeches", "corpus", function( 
   .Object,
   s_attribute_date = grep("date", s_attributes(.Object), value = TRUE),
   s_attribute_name = grep("name", s_attributes(.Object), value = TRUE),
   gap = 500, mc = FALSE, verbose = TRUE, progress = TRUE
 ){
-  m <- make_region_matrix(.Object, s_attribute = s_attribute_name) # in polmineR, not exported
-  dates <- s_attributes(.Object, s_attribute = s_attribute_date, unique = FALSE)
-  strucs <- 0L:(nrow(m) - 1L)
-  speakers <- s_attributes(.Object, s_attribute = s_attribute_name, unique = FALSE)
   
-  if (length(dates) != length(speakers))
-    stop(
-      sprintf(
-        "Number of regions for s_attribute '%s' and s_attribute '%s' not identical - procedure will not work",
-        s_attribute_date, s_attribute_name)
-    )
+  speakers <- s_attributes(.Object, s_attribute = s_attribute_name, unique = FALSE)
+  regions <- s_attr_regions(
+    corpus = .Object@corpus, s_attr = s_attribute_name,
+    registry = .Object@registry_dir, data_dir = .Object@data_dir
+  )
+  strucs <- 0L:(nrow(regions) - 1L)
+  
+  
+  sibling <- s_attr_is_sibling(
+    x = s_attribute_date, y = s_attribute_name,
+    corpus = .Object@corpus, registry = .Object@registry_dir
+  )
 
-  chunks_cpos <- split(x = m, f = speakers)
+  if (isTRUE(sibling)){
+    dates <- s_attributes(.Object, s_attribute = s_attribute_date, unique = FALSE)
+  } else {
+    # This additional check is not strictly necessary - but we are on the safe side
+    descendent <- s_attr_is_descendent(
+      x = s_attribute_name, y = s_attribute_date,
+      corpus = .Object@corpus, registry = .Object@registry_dir
+    )
+    if (isFALSE(descendent))
+      stop(
+        "Unknown scenario: s_attribute_name and s_attribute_date are not ",
+        "siblings, but s_attribute_name is not a descendent of s_attribute_date"
+      )
+    s <- cpos2struc(.Object, s_attr = s_attribute_date, cpos = regions[,1L])
+    dates <- struc2str(.Object, s_attr = s_attribute_date, struc = s)
+  }
+
+  chunks_cpos <- split(x = regions, f = speakers)
   chunks_dates <- split(x = dates, f = speakers)
   chunks_strucs <- split(x = strucs, f = speakers)
 
@@ -207,22 +229,31 @@ setMethod("as.speeches", "corpus", function(
       )))
     }
     
-    distance <- mx[,1][2L:nrow(mx)] - mx[,2][1L:(nrow(mx) - 1L)]
+    distance <- mx[,1L][2L:nrow(mx)] - mx[,2L][1L:(nrow(mx) - 1L)]
     beginning <- c(TRUE, ifelse(distance > gap, TRUE, FALSE))
+    
+    # new speech begins also if date of region is not identical with date of 
+    # preceding region
     beginning <- ifelse(
-      c(TRUE, chunks_dates[[i]][2L:length(chunks_dates[[i]])] == chunks_dates[[i]][1L:(length(chunks_dates[[i]]) - 1L)]),
+      c(
+        TRUE,
+        chunks_dates[[i]][2L:length(chunks_dates[[i]])] == chunks_dates[[i]][1L:(length(chunks_dates[[i]]) - 1L)]
+      ),
       beginning,
       TRUE
     )
     razor <- cumsum(beginning)
     vec_dates <- chunks_dates[[i]][beginning]
     
-    # The speech_no vector indicates the number of the speech at a date for the speeches
-    # referred to with the vec_dates vector. These lines have seen some revisions to 
-    # make the procedure robust. Resorting to the for loop is able to handle situations
-    # robustly when dates are not increasing throughout.
+    # The speech_no vector indicates the number of the speech at a date for the
+    # speeches referred to with the vec_dates vector. These lines have seen some
+    # revisions to make the procedure robust. Resorting to the for loop is able
+    # to handle situations robustly when dates are not increasing throughout.
     unique_dates <- unique(vec_dates)
-    speech_no_aux <- setNames(rep(1L, times = length(unique_dates)), nm = unique_dates)
+    speech_no_aux <- setNames(
+      rep(1L, times = length(unique_dates)),
+      nm = unique_dates
+    )
     speech_no <- rep(NA, times = length(vec_dates))
     for (k in seq_along(vec_dates)){
       speech_no[k] <- speech_no_aux[[vec_dates[k]]]
@@ -255,6 +286,7 @@ setMethod("as.speeches", "corpus", function(
       }
     )
   }
+  
   y <- if (progress){
     pblapply(seq_along(chunks_cpos), .iter_fn, cl = mc)
   } else {